Fixed the half-pixel bias in the Fourier Gaussian registration filter.
-- Centered the Fourier Gaussian smoothing kernel on the sample ifftshift moves to the origin, which makes the filter zero-phase and holds the correlation peaks it smooths in place on even-sized frames.
-- Removed the systematic one-pixel-per-axis offset the bias placed on every even-sized frame, which the iterative reference build compounded into a several-pixel shift that narrowed the registered crop.
-- Added a regression test asserting that the filter carries no phase on even dimensions.
-- Updated the two-step refinement and minimum-frame-count metric tests to the bounds the corrected offsets support.
-- Demoted mean_centered_meshgrid to a module-private helper of detection.utils, its one remaining consumer.
-- Added a ruff S603 suppression to the GUI viewer subprocess launch, whose command resolves from the interpreter path and runs without a shell.
-- Fixed two mypy errors in detect_rois.py by returning the split variance ratio as a Python float and clamping the weight threshold with a float32 zero.
-- Updated the project version to 2.0.0rc5 and re-exported the development environment specification.

+++ src/cindra/detection/detect_rois.py
@@ -545,7 +545,7 @@ def _check_split_components(
     best_component = np.argmax(component_variances)
     residual_energy = np.dot(pixel_frames.ravel(), pixel_frames.ravel())
     variance_ratio = (total_energy - residual_energy) / single_component_variance
-    return variance_ratio, (
+    return float(variance_ratio), (
         component_masks[best_component],
         projections[best_component],
         component_frames[best_component],
@@ -751,7 +751,7 @@ def _extend_iteratively(
         # a fraction of the peak to discard weakly correlated boundary pixels.
         flat_indices = y_pixels * width + x_pixels
         weights = frames[np.ix_(active_frame_indices, flat_indices)].mean(axis=0)
-        active_mask = weights > max(0, weights.max() * _MINIMUM_WEIGHT_FRACTION)
+        active_mask = weights > max(np.float32(0), weights.max() * _MINIMUM_WEIGHT_FRACTION)
         active_count = active_mask.sum()
         if active_count == 0:
             break
